requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Babel==3.1.0
Flask-Limiter==3.3.1
Flask-Talisman==1.0.0
Flask-APScheduler==1.12.4
Flask-Mail==0.9.1
python-dotenv==1.0.0
email-validator==2.0.0
phonenumbers==8.13.17
reportlab==4.0.4
redis==4.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
bcrypt==4.0.1
pytest==7.4.0